feat(parse): show word definition in GNT passage parse feedback
Extracts GNTFeedback into its own component (GNTFeedback.tsx) and adds
the lemma/gloss/frequency row that was missing from the GNT passage
parsing challenge — matching the behaviour already present in the
standard verb parse quiz.

closes #104
